--- BIP39.NET.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BIP39.NET", "BIP39.NET.csproj", "{517A4285-49E4-4478-9388-DCF497EEA68A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{517A4285-49E4-4478-9388-DCF497EEA68A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{517A4285-49E4-4478-9388-DCF497EEA68A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{517A4285-49E4-4478-9388-DCF497EEA68A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{517A4285-49E4-4478-9388-DCF497EEA68A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
