+++ README.md
@@ -1,4 +1,4 @@
-# BIP39.NET-Portable
+# BIP39.NETCORE
 BIP39 portable class library written in c#, original author thashiznets@yahoo.com.au
 
-NuGet package https://www.nuget.org/packages/BIP39/
+NuGet package https://www.nuget.org/packages/BIP39.NETCORE/
