+++ src/Bitcoin.BIP39/Wordlists/ChineseTraditional.cs
