+++ src/Bitcoin.BIP39/Wordlists/English.cs
