+++ src/Bitcoin.BIP39/Wordlists/Japanese.cs
