+++ src/Bitcoin.BIP39/Wordlists/Spanish.cs
